Fixed compiler warnings

+++ App/Source/UI/AudioDebugger.cpp
@@ -21,28 +21,28 @@ namespace App
 		bool masterEnabled = m_GameBoy->m_MMU.ReadRegisterBit(Core::HW_NR52_SOUND_TOGGLE, Core::NR52_AUDIO_ON); ImGui::Checkbox("Enabled", &masterEnabled);
 
 		std::vector<float> masterBuffer = m_GameBoy->m_APU.GetMasterAudioBuffer();
-		ImGui::PlotLines("Master", masterBuffer.data(), masterBuffer.size(), 0, nullptr, -1.0f, 1.0f, ImVec2(0, 40.0f));
+		ImGui::PlotLines("Master", masterBuffer.data(), static_cast<int>(masterBuffer.size()), 0, nullptr, -1.0f, 1.0f, ImVec2(0, 40.0f));
 
 
 		ImGui::SeparatorText("Channel 1 - Square");
 		bool ch1Enabled = m_GameBoy->m_MMU.ReadRegisterBit(Core::HW_NR52_SOUND_TOGGLE, Core::NR52_CH1_ON); ImGui::Checkbox("Enabled", &ch1Enabled);
 
 		std::vector<float> ch1Buffer = m_GameBoy->m_APU.GetCh1AudioBuffer();
-		ImGui::PlotLines("CH1 - Square", ch1Buffer.data(), ch1Buffer.size(), 0, nullptr, -1.0f, 1.0f, ImVec2(0, 40.0f));
+		ImGui::PlotLines("CH1 - Square", ch1Buffer.data(), static_cast<int>(ch1Buffer.size()), 0, nullptr, -1.0f, 1.0f, ImVec2(0, 40.0f));
 
 
 		ImGui::SeparatorText("Channel 2 - Square");
 		bool ch2Enabled = m_GameBoy->m_MMU.ReadRegisterBit(Core::HW_NR52_SOUND_TOGGLE, Core::NR52_CH2_ON); ImGui::Checkbox("Enabled", &ch2Enabled);
 
 		std::vector<float> ch2Buffer = m_GameBoy->m_APU.GetCh2AudioBuffer();
-		ImGui::PlotLines("CH2 - Square", ch2Buffer.data(), ch2Buffer.size(), 0, nullptr, -1.0f, 1.0f, ImVec2(0, 40.0f));
+		ImGui::PlotLines("CH2 - Square", ch2Buffer.data(), static_cast<int>(ch2Buffer.size()), 0, nullptr, -1.0f, 1.0f, ImVec2(0, 40.0f));
 
 
 		ImGui::SeparatorText("Channel 3 - Wave");
 		bool ch3Enabled = m_GameBoy->m_MMU.ReadRegisterBit(Core::HW_NR52_SOUND_TOGGLE, Core::NR52_CH3_ON); ImGui::Checkbox("Enabled", &ch3Enabled);
 
 		std::vector<float> ch3Buffer = m_GameBoy->m_APU.GetCh3AudioBuffer();
-		ImGui::PlotLines("CH3 - Wave", ch3Buffer.data(), ch3Buffer.size(), 0, nullptr, -1.0f, 1.0f, ImVec2(0, 40.0f));
+		ImGui::PlotLines("CH3 - Wave", ch3Buffer.data(), static_cast<int>(ch3Buffer.size()), 0, nullptr, -1.0f, 1.0f, ImVec2(0, 40.0f));
 
 		// TODO: Need to fix this abomination
 		float arr[] =
@@ -73,7 +73,7 @@ namespace App
 		bool ch4Enabled = m_GameBoy->m_MMU.ReadRegisterBit(Core::HW_NR52_SOUND_TOGGLE, Core::NR52_CH4_ON); ImGui::Checkbox("Enabled", &ch4Enabled);
 
 		std::vector<float> ch4Buffer = m_GameBoy->m_APU.GetCh4AudioBuffer();
-		ImGui::PlotLines("CH4 - Noise", ch4Buffer.data(), ch4Buffer.size(), 0, nullptr, -1.0f, 1.0f, ImVec2(0, 40.0f));
+		ImGui::PlotLines("CH4 - Noise", ch4Buffer.data(), static_cast<int>(ch4Buffer.size()), 0, nullptr, -1.0f, 1.0f, ImVec2(0, 40.0f));
 
 		ImGui::SeparatorText("Registers");
